Address comments from a review
