+++ packages/app/features/dashboard/beat/ActiveOffer.tsx
@@ -1,14 +1,14 @@
 import { Text, View } from "app/design-system";
-import { OfferRecord } from "app/types";
-import { CancelOfferBtn } from "app/ui/buttons/CancelOfferBtn";
+import { Offer } from "app/types";
+import { CancelOfferBtn } from "app/features/dashboard/beat/CancelOfferBtn";
 import { Entry, EntryHolder } from "app/api/graphql";
-import { CreateOfferBtn } from "app/ui/buttons/CreateOfferBtn";
+import { CreateOfferBtn } from "app/features/dashboard/beat/CreateOfferBtn";
 import { LinkWithLabel } from "app/ui/links/LinkWithLabel";
 
 type ActiveOfferProps = {
   entry: Entry;
   index: number;
-  offer: OfferRecord;
+  offer: Offer;
   holders?: EntryHolder[] | null;
 };
 
@@ -27,7 +27,7 @@ export function ActiveOffer({
 
   return (
     <View className={`md:flex-row py-3 px-5 md:items-center ${bgColor}`}>
-      <LinkWithLabel type="offer" id={id} />
+      <LinkWithLabel type="offer" id={id.toString()} />
 
       <View className="md:flex-1 my-2 md:mx-2 md:my-0">
         <View className="md:flex-col">
@@ -37,8 +37,8 @@ export function ActiveOffer({
           </Text>
         </View>
       </View>
-      <CreateOfferBtn entry={entry} holders={holders} offerId={id} />
-      <CancelOfferBtn entry={entry} offerId={id} />
+      <CreateOfferBtn entry={entry} holders={holders} offerId={id.toString()} />
+      <CancelOfferBtn entry={entry} offerId={id.toString()} />
     </View>
   );
 }
