+++ packages/app/features/dashboard/beat/OwnerOffers.tsx
@@ -1,10 +1,9 @@
 import { CollapsableView } from "app/ui/CollapsableView";
 import { View, Text } from "app/design-system";
 import { ArrowsUpDownIcon } from "app/ui/icons/arrows-up-down";
-import { OfferRecord } from "app/types";
 import { ActiveOffer } from "./ActiveOffer";
 import { Entry, EntryHolder } from "app/api/graphql";
-import { CreateOfferBtn } from "app/ui/buttons/CreateOfferBtn";
+import { CreateOfferBtn } from "app/features/dashboard/beat/CreateOfferBtn";
 import { useUserOffers } from "app/hooks/useUserOffers";
 import { useRecoilValue } from "recoil";
 import { userAtom } from "app/state/user";
@@ -14,26 +13,11 @@ type OwnerOffersProps = {
   holders?: EntryHolder[] | null;
 };
 
-type ListOffersProps = {
-  offers: OfferRecord[];
-  entry: Entry;
-  holders?: EntryHolder[] | null;
-};
-
-type CreateOfferProps = {
-  entry: Entry;
-  holders?: EntryHolder[] | null;
-};
-
 export function OwnerOffers({ entry, holders }: OwnerOffersProps) {
   const user = useRecoilValue(userAtom);
-  const { offers }: { offers: OfferRecord[] } = useUserOffers(
-    user?.publicKey,
-    entry.issuer,
-    entry.code
-  );
+  const { offers } = useUserOffers(user?.publicKey, entry.issuer, entry.code);
 
-  const Offers = ({ offers, entry }: ListOffersProps) => {
+  const Offers = () => {
     return (
       <CollapsableView headerText="Your active offers" icon={ArrowsUpDownIcon}>
         <View>
@@ -51,7 +35,7 @@ export function OwnerOffers({ entry, holders }: OwnerOffersProps) {
     );
   };
 
-  const CreateOfferRow = ({ entry, holders }: CreateOfferProps) => {
+  const CreateOfferRow = () => {
     return (
       <CollapsableView headerText="Your active offers" icon={ArrowsUpDownIcon}>
         <View>
@@ -66,13 +50,5 @@ export function OwnerOffers({ entry, holders }: OwnerOffersProps) {
     );
   };
 
-  return (
-    <>
-      {offers && offers.length > 0 ? (
-        <Offers offers={offers} entry={entry} holders={holders} />
-      ) : (
-        <CreateOfferRow entry={entry} holders={holders} />
-      )}
-    </>
-  );
+  return <>{offers && offers.length > 0 ? <Offers /> : <CreateOfferRow />}</>;
 }
